refactor: change entity table define to offcial docs
